+++ src/components/swap/FeesAndRateDetails.tsx
@@ -45,6 +45,13 @@ export const FeesAndRateDetails = () => {
   const protocolFee = useMemo(() => getProtocolFee(fees), [fees]);
   const totalCost = protocolFee + networkFeesValue;
 
+  const priceImpact = useMemo(() => {
+    if (!tokenPrices) return 0;
+    const input = Number(tokenPrices.input);
+    const output = Number(tokenPrices.output);
+    return (1 - (output - protocolFee) / input) * 100;
+  }, [tokenPrices, protocolFee]);
+
   const isBitcoinChains = outputAsset?.symbol.includes(BTC.symbol);
   const formattedRate = useMemo(
     () => Number(rate.toFixed(isBitcoinChains ? 7 : 3)),
@@ -133,6 +140,7 @@ export const FeesAndRateDetails = () => {
                         networkFee={networkFeesValue}
                         protocolFee={protocolFee}
                         rate={formattedRate}
+                        priceImpact={priceImpact}
                       />
                     </TooltipWrapper>
                   )}